requirements.txt
flask=3.1.2
mysql-connector-python=9.5.0